Analytics tracking runs without re-rendering Right after giving consent Without useEffect On Click on Anchors aswell
