+++ pages/index.tsx
@@ -1,13 +1,7 @@
-import { useEffect, useState } from 'react'
 import { domain } from 'lib/config'
 import { resolveNotionPage } from 'lib/resolve-notion-page'
 import { NotionPage } from 'components'
-
-import { useEvents } from 'lib/useEvents'
 import { PageProps } from 'lib/types'
-
-import { CookieBanner } from '@palmabit/react-cookie-law'
-import { bannerStyles } from '../styles/banner'
 
 export const getStaticProps = async () => {
   try {
@@ -24,21 +18,5 @@ export const getStaticProps = async () => {
 }
 
 export default function NotionDomainPage(props: PageProps) {
-  const { cookieConsent, setCookieConsent } = useEvents()
-
-  return (
-    <>
-      <NotionPage {...props} />
-      {!cookieConsent && (
-        <CookieBanner
-          message='Wir verwenden kleine Kekse - auch genannt "Cookies". Willst du die auch?'
-          wholeDomain={true}
-          onAccept={() => setTimeout(() => setCookieConsent(true), 200)}
-          acceptButtonText='Ja, klar!'
-          managePreferencesButtonText='Einstellungen'
-          styles={bannerStyles}
-        />
-      )}
-    </>
-  )
+  return <NotionPage {...props} />
 }
